requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
sentence-transformers
scikit-learn
requests
prophet
python-dotenv
